requirements.txt
astroid==2.15.6
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
dill==0.3.7
idna==3.4
isort==5.12.0
jira==3.5.2
lazy-object-proxy==1.9.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
platformdirs==3.10.0
pydash==7.0.6
pylint==2.17.5
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
urllib3==2.0.4
wrapt==1.15.0
